Fixed #87
