+++ packages/explorer/src/components/BondForm/props.js
@@ -1,0 +1,15 @@
+// @flow
+export type BondFormProps = {
+  bondedAmount: string,
+  delegateAddress: string,
+  handleSubmit: any => void,
+  loading: boolean,
+  onCancel: any => void,
+  pristine: boolean,
+  tokenBalance: string,
+  valid: boolean,
+  values: {
+    amount: string,
+    to: string,
+  },
+}
