+++ packages/explorer/src/components/CTABanner.js
@@ -1,6 +1,5 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import Button from './Button'
 
 const CTABanner = styled(
   class CTABanner extends React.Component {
