+++ packages/explorer/src/components/InlineAccount/props.js
@@ -1,0 +1,5 @@
+// @flow
+export type InlineAccountProps = {
+  address: string,
+  truncate: number,
+}
