+++ packages/explorer/src/components/ScrollToTopOnMount.js
@@ -4,6 +4,7 @@ class ScrollToTopOnMount extends React.Component {
   componentDidMount() {
     window.scrollTo(0, 0)
   }
+
   render() {
     return null
   }
