+++ packages/explorer/src/components/TranscoderStat/props.js
@@ -1,0 +1,10 @@
+// @flow
+export type TranscoderStatProps = {|
+  decimals?: number,
+  label: string,
+  symbol?: string,
+  type?: 'percentage' | 'token',
+  unit?: string,
+  value: string | number,
+  width: string,
+|}
